Move package to nodeconf
